optimize code
